+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Added
 
 + You can now pull CompatabilityFramework into a Home Brew load order without seeing classes that aren't meant to be in Home Brew
-+ Implemented the Dynamic ASI +2 patch that will add +2 ASI to every feat that doesn't already have it (except any feat that provides an option to select from Home Brew ASI passive list)
++ Implemented the Dynamic ASI +2 patch that will add +2 ASI to every feat that doesn't already have it 
+(except any feat that provides an option to select from Home Brew ASI passive list)
++ Implemented the Dynamic Longbow Patch. Every longbow in the game will be patched to use Strength as an attack roll attribute
+if the character's Strength is greater than their Dexterity
 
 ## [0.0.1] - <YYYY-MM-DD>